requirements.txt
argparse
psutil
dgl
torch
torchvision # (Just for tensorboard logging)
pytorch-ignite
matplotlib
scikitlearn